Upgraded libraries and split executor files up
