+++ .python-version
@@ -1,1 +1,1 @@
-3.13
+3.14
