+++ src/mini/__init__.py
@@ -1,9 +1,17 @@
+from mini.executor import Executor, ProgressDisplay, get_progress
 from mini.experiment import Experiment
 from mini.hither import AsyncCallback, AsyncBatchCallback, Callback
+from mini.local_executor import LocalExecutor
+from mini.modal_executor import ModalExecutor
 
 __all__ = [
     'AsyncBatchCallback',
     'AsyncCallback',
     'Callback',
+    'Executor',
     'Experiment',
+    'LocalExecutor',
+    'ModalExecutor',
+    'ProgressDisplay',
+    'get_progress',
 ]
