+++ src/mini/executor.py
@@ -1,24 +1,12 @@
 """
-Executor-based experiment runner.
+Executor protocol and progress reporting infrastructure.
 
-Provides a simple ``map``-style interface for running experiment sweeps,
-similar to ``concurrent.futures.ThreadPoolExecutor``.  Functions are passed
-in (not decorated), so you can switch between local and remote execution
-without changing the experiment code.
+Provides shared components for executor implementations:
+- ``Executor`` protocol defining the map interface
+- ``ProgressDisplay`` for tracking concurrent job progress
+- ``get_progress()`` for mapped functions to report progress
 
 Example::
-
-    executor = LocalExecutor("my-experiment", max_workers=2)
-    results = list(executor.map(train, [{"seed": 1}, {"seed": 2}]))
-
-    # Switch to Modal for scale
-    app = modal.App("my-experiment")
-    executor = ModalExecutor(app).with_modal_kwargs(gpu="T4")
-    results = list(executor.map(train, [{"seed": i} for i in range(10)]))
-
-Progress reporting
-------------------
-Mapped functions can optionally report fine-grained progress::
 
     from mini.executor import get_progress
 
@@ -31,33 +19,20 @@ Mapped functions can optionally report fine-grained progress::
             if progress:
                 progress.update(1, message=f"loss={loss:.4f}")
         return result
-
-``get_progress()`` returns a ``JobProgress`` handle when running inside an
-executor, or ``None`` otherwise.  This works for ``LocalExecutor`` (via a
-context variable set per-thread).  For ``ModalExecutor``, fine-grained
-progress is visible through Modal's built-in log streaming (functions can
-simply ``print``); job-level completion is tracked locally.
 """
 
 from __future__ import annotations
 
 import contextvars
-import logging
 import sys
 import threading
 import time
-from concurrent.futures import Future, ThreadPoolExecutor
 from typing import Any, Callable, Iterable, Iterator, Protocol, TypeVar
 
-log = logging.getLogger(__name__)
-
-T = TypeVar('T')
 R = TypeVar('R')
 
 __all__ = [
     'Executor',
-    'LocalExecutor',
-    'ModalExecutor',
     'JobProgress',
     'ProgressDisplay',
     'get_progress',
@@ -198,7 +173,12 @@ def get_progress() -> JobProgress | None:
 class Executor(Protocol):
     """Protocol for running a function over a sweep of inputs."""
 
-    def map(self, fn: Callable[..., R], *iterables: Iterable[Any], kwargs: dict[str, Any] | None = None) -> Iterator[R]:
+    def map(
+        self,
+        fn: Callable[..., R],
+        *iterables: Iterable[Any],
+        kwargs: dict[str, Any] | None = None,
+    ) -> Iterator[R]:
         """
         Map *fn* over one or more iterables.
 
@@ -214,104 +194,3 @@ class Executor(Protocol):
             executor.map(fn, [1, 2], kwargs={'k': 'v'})     # fn(1, k='v'), fn(2, k='v')
         """
         ...
-
-
-# ---------------------------------------------------------------------------
-# LocalExecutor
-# ---------------------------------------------------------------------------
-
-
-class LocalExecutor:
-    """
-    Run functions locally using a thread pool.
-
-    Progress is tracked per-job via ``get_progress()`` inside mapped
-    functions, and a live status line is rendered to stderr.
-    """
-
-    def __init__(self, name: str, max_workers: int = 1):
-        self.name = name
-        self.max_workers = max_workers
-
-    def map(self, fn: Callable[..., R], *iterables: Iterable[Any], kwargs: dict[str, Any] | None = None) -> Iterator[R]:
-        kw = kwargs or {}
-        args_list = list(zip(*iterables, strict=False))
-        n = len(args_list)
-        if n == 0:
-            return
-
-        log.info('[%s] Running %d jobs with %d workers', self.name, n, self.max_workers)
-        display = ProgressDisplay(n)
-
-        def run_one(index: int, args: tuple) -> R:
-            progress = display.job_started(index)
-            token = _current_progress.set(progress)
-            try:
-                result = fn(*args, **kw)
-                display.job_completed(index)
-                return result
-            except Exception as e:
-                display.job_failed(index, str(e)[:80])
-                raise
-            finally:
-                _current_progress.reset(token)
-
-        with ThreadPoolExecutor(max_workers=self.max_workers) as pool:
-            futures: list[Future[R]] = [pool.submit(run_one, i, args) for i, args in enumerate(args_list)]
-            try:
-                for future in futures:
-                    yield future.result()
-            finally:
-                display.finish()
-
-
-# ---------------------------------------------------------------------------
-# ModalExecutor
-# ---------------------------------------------------------------------------
-
-
-class ModalExecutor:
-    """
-    Run functions on Modal.
-
-    Fine-grained progress (epoch-level etc.) is visible through Modal's
-    built-in log streaming --- mapped functions can simply ``print()``.
-    Job-level completion is tracked locally via a progress display.
-
-    Usage::
-
-        app = modal.App("my-experiment")
-        executor = ModalExecutor(app).with_modal_kwargs(gpu="T4", timeout=3600)
-        results = list(executor.map(train, configs))
-    """
-
-    def __init__(self, app: Any, modal_fn_kwargs: dict[str, Any] | None = None):
-        self.app = app
-        self.modal_fn_kwargs: dict[str, Any] = modal_fn_kwargs or {}
-
-    def with_modal_kwargs(self, **kwargs: Any) -> ModalExecutor:
-        """Return a new executor with additional Modal function kwargs merged in."""
-        return ModalExecutor(self.app, {**self.modal_fn_kwargs, **kwargs})
-
-    def map(self, fn: Callable[..., R], *iterables: Iterable[Any], kwargs: dict[str, Any] | None = None) -> Iterator[R]:
-        iterables_lists: list[list] = [list(it) for it in iterables]
-        n = len(iterables_lists[0]) if iterables_lists else 0
-        if n == 0:
-            return
-
-        log.info('[ModalExecutor] Running %d jobs on Modal', n)
-
-        # Wrap fn as a Modal function.  The decorator must be applied
-        # *before* app.run() starts the app.
-        modal_fn = self.app.function(**self.modal_fn_kwargs)(fn)
-
-        display = ProgressDisplay(n)
-        # We don't have per-step progress from Modal, but we can track
-        # job-level completion as results come back.
-        with self.app.run():
-            try:
-                for i, result in enumerate(modal_fn.map(*iterables_lists, kwargs=kwargs or {})):
-                    display.job_completed(i)
-                    yield result
-            finally:
-                display.finish()
