+++ tests/mini/test_executor.py
@@ -5,7 +5,9 @@ import time
 
 import pytest
 
-from mini.executor import LocalExecutor, ModalExecutor, ProgressDisplay, get_progress
+from mini.executor import ProgressDisplay, get_progress
+from mini.local_executor import LocalExecutor
+from mini.modal_executor import ModalExecutor
 
 
 # ---------------------------------------------------------------------------
